+++ src/terminal.go
@@ -164,6 +164,7 @@ type eachLine struct {
 }
 
 type itemLine struct {
+	valid     bool
 	firstLine int
 	numLines  int
 	cy        int
@@ -179,11 +180,15 @@ type itemLine struct {
 }
 
 func (t *Terminal) markEmptyLine(line int) {
-	t.prevLines[line] = itemLine{firstLine: line, empty: true}
+	if t.window != t.inputWindow {
+		t.prevLines[line] = itemLine{valid: true, firstLine: line, empty: true}
+	}
 }
 
 func (t *Terminal) markOtherLine(line int) {
-	t.prevLines[line] = itemLine{firstLine: line, other: true}
+	if t.window != t.inputWindow {
+		t.prevLines[line] = itemLine{valid: true, firstLine: line, other: true}
+	}
 }
 
 type fitpad struct {
@@ -241,6 +246,9 @@ type Terminal struct {
 	previewLabel       labelPrinter
 	previewLabelLen    int
 	previewLabelOpts   labelOpts
+	inputLabel         labelPrinter
+	inputLabelLen      int
+	inputLabelOpts     labelOpts
 	pointer            string
 	pointerLen         int
 	pointerEmpty       string
@@ -297,11 +305,19 @@ type Terminal struct {
 	listener           net.Listener
 	listenUnsafe       bool
 	borderShape        tui.BorderShape
+	listBorderShape    tui.BorderShape
+	inputBorderShape   tui.BorderShape
+	listLabel          labelPrinter
+	listLabelLen       int
+	listLabelOpts      labelOpts
 	cleanExit          bool
 	executor           *util.Executor
 	paused             bool
 	border             tui.Window
 	window             tui.Window
+	inputWindow        tui.Window
+	inputBorder        tui.Window
+	wborder            tui.Window
 	pborder            tui.Window
 	pwindow            tui.Window
 	borderWidth        int
@@ -389,6 +405,8 @@ const (
 	reqReinit
 	reqFullRedraw
 	reqResize
+	reqRedrawInputLabel
+	reqRedrawListLabel
 	reqRedrawBorderLabel
 	reqRedrawPreviewLabel
 	reqClose
@@ -429,6 +447,8 @@ const (
 	actBackwardWord
 	actCancel
 	actChangeBorderLabel
+	actChangeListLabel
+	actChangeInputLabel
 	actChangeHeader
 	actChangeMulti
 	actChangePreviewLabel
@@ -463,6 +483,8 @@ const (
 	actToggleTrackCurrent
 	actToggleHeader
 	actToggleWrap
+	actToggleMultiLine
+	actToggleHscroll
 	actTrackCurrent
 	actUntrackCurrent
 	actDown
@@ -487,6 +509,8 @@ const (
 	actTogglePreviewWrap
 	actTransform
 	actTransformBorderLabel
+	actTransformListLabel
+	actTransformInputLabel
 	actTransformHeader
 	actTransformPreviewLabel
 	actTransformPrompt
@@ -821,11 +845,17 @@ func NewTerminal(opts *Options, eventBox *util.EventBox, executor *util.Executor
 		listenAddr:         opts.ListenAddr,
 		listenUnsafe:       opts.Unsafe,
 		borderShape:        opts.BorderShape,
+		listBorderShape:    opts.ListBorderShape,
+		inputBorderShape:   opts.InputBorderShape,
 		borderWidth:        1,
+		listLabel:          nil,
+		listLabelOpts:      opts.ListLabel,
 		borderLabel:        nil,
 		borderLabelOpts:    opts.BorderLabel,
 		previewLabel:       nil,
 		previewLabelOpts:   opts.PreviewLabel,
+		inputLabel:         nil,
+		inputLabelOpts:     opts.InputLabel,
 		cleanExit:          opts.ClearOnExit,
 		executor:           executor,
 		paused:             opts.Phony,
@@ -878,13 +908,21 @@ func NewTerminal(opts *Options, eventBox *util.EventBox, executor *util.Executor
 		executing:          util.NewAtomicBool(false),
 		lastAction:         actStart,
 		lastFocus:          minItem.Index()}
+
+	// This should be called before accessing tui.Color*
+	tui.InitTheme(opts.Theme, renderer.DefaultTheme(), opts.Black)
+
 	t.prompt, t.promptLen = t.parsePrompt(opts.Prompt)
 	// Pre-calculated empty pointer and marker signs
 	t.pointerEmpty = strings.Repeat(" ", t.pointerLen)
 	t.markerEmpty = strings.Repeat(" ", t.markerLen)
+	t.listLabel, t.listLabelLen = t.ansiLabelPrinter(opts.ListLabel.label, &tui.ColListLabel, false)
 	t.borderLabel, t.borderLabelLen = t.ansiLabelPrinter(opts.BorderLabel.label, &tui.ColBorderLabel, false)
 	t.previewLabel, t.previewLabelLen = t.ansiLabelPrinter(opts.PreviewLabel.label, &tui.ColPreviewLabel, false)
-	if opts.Separator == nil || len(*opts.Separator) > 0 {
+	t.inputLabel, t.inputLabelLen = t.ansiLabelPrinter(opts.InputLabel.label, &tui.ColInputLabel, false)
+
+	// Disable separator by default if input border is set
+	if opts.Separator == nil && !t.inputBorderShape.Visible() || opts.Separator != nil && len(*opts.Separator) > 0 {
 		bar := "─"
 		if opts.Separator != nil {
 			bar = *opts.Separator
@@ -971,6 +1009,7 @@ func (t *Terminal) environ() []string {
 	env = append(env, "FZF_PROMPT="+string(t.promptString))
 	env = append(env, "FZF_PREVIEW_LABEL="+t.previewLabelOpts.label)
 	env = append(env, "FZF_BORDER_LABEL="+t.borderLabelOpts.label)
+	env = append(env, "FZF_LIST_LABEL="+t.listLabelOpts.label)
 	env = append(env, fmt.Sprintf("FZF_TOTAL_COUNT=%d", t.count))
 	env = append(env, fmt.Sprintf("FZF_MATCH_COUNT=%d", t.merger.Length()))
 	env = append(env, fmt.Sprintf("FZF_SELECT_COUNT=%d", len(t.selected)))
@@ -1101,6 +1140,17 @@ func (t *Terminal) ansiLabelPrinter(str string, color *tui.ColorPair, fill bool)
 	return printFn, length
 }
 
+// Temporarily switch 'window' so that we can use the existing windows with
+// a different window
+func (t *Terminal) withInputWindow(f func()) {
+	prevWindow := t.window
+	if t.inputWindow != nil {
+		t.window = t.inputWindow
+	}
+	f()
+	t.window = prevWindow
+}
+
 func (t *Terminal) parsePrompt(prompt string) (func(), int) {
 	var state *ansiState
 	prompt = firstLine(prompt)
@@ -1112,7 +1162,7 @@ func (t *Terminal) parsePrompt(prompt string) (func(), int) {
 	//             // unless the part has a non-default ANSI state
 	loc := whiteSuffix.FindStringIndex(trimmed)
 	if loc != nil {
-		blankState := ansiOffset{[2]int32{int32(loc[0]), int32(loc[1])}, ansiState{-1, -1, tui.AttrClear, -1, nil}}
+		blankState := ansiOffset{[2]int32{int32(loc[0]), int32(loc[1])}, ansiState{tui.ColPrompt.Fg(), tui.ColPrompt.Bg(), tui.AttrClear, -1, nil}}
 		if item.colors != nil {
 			lastColor := (*item.colors)[len(*item.colors)-1]
 			if lastColor.offset[1] < int32(loc[1]) {
@@ -1126,11 +1176,13 @@ func (t *Terminal) parsePrompt(prompt string) (func(), int) {
 		}
 	}
 	output := func() {
-		line := t.promptLine()
 		wrap := t.wrap
 		t.wrap = false
-		t.printHighlighted(
-			Result{item: item}, tui.ColPrompt, tui.ColPrompt, false, false, line, line, true, nil, nil)
+		t.withInputWindow(func() {
+			line := t.promptLine()
+			t.printHighlighted(
+				Result{item: item}, tui.ColPrompt, tui.ColPrompt, false, false, line, line, true, nil, nil)
+		})
 		t.wrap = wrap
 	}
 	_, promptLen := t.processTabs([]rune(trimmed), 0)
@@ -1544,6 +1596,15 @@ func (t *Terminal) resizeWindows(forcePreview bool, redrawBorder bool) {
 	if t.window != nil {
 		t.window = nil
 	}
+	if t.wborder != nil {
+		t.wborder = nil
+	}
+	if t.inputWindow != nil {
+		t.inputWindow = nil
+	}
+	if t.inputBorder != nil {
+		t.inputBorder = nil
+	}
 	if t.pborder != nil {
 		t.pborder = nil
 	}
@@ -1570,10 +1631,10 @@ func (t *Terminal) resizeWindows(forcePreview bool, redrawBorder bool) {
 		offsets[1] -= 1 + bw
 		offsets[2] += 1 + bw
 	}
-	if t.border == nil && t.borderShape != tui.BorderNone {
+	if t.border == nil && t.borderShape.Visible() {
 		t.border = t.tui.NewWindow(
 			marginInt[0]+offsets[0], marginInt[3]+offsets[1], width+offsets[2], height+offsets[3],
-			false, tui.MakeBorderStyle(t.borderShape, t.unicode))
+			tui.WindowBase, tui.MakeBorderStyle(t.borderShape, t.unicode), true)
 	}
 
 	// Add padding to margin
@@ -1583,6 +1644,54 @@ func (t *Terminal) resizeWindows(forcePreview bool, redrawBorder bool) {
 	width -= paddingInt[1] + paddingInt[3]
 	height -= paddingInt[0] + paddingInt[2]
 
+	// Adjust position and size of the list window if input border is set
+	inputWindowHeight := 0
+	inputBorderHeight := 0
+	shift := 0
+	shrink := 0
+	hasInputWindow := t.inputBorderShape.Visible()
+	if hasInputWindow {
+		inputWindowHeight = 2
+		if t.noSeparatorLine() {
+			inputWindowHeight--
+		}
+		inputBorderHeight = borderLines(t.inputBorderShape) + inputWindowHeight
+		if t.layout == layoutReverse {
+			shift = inputBorderHeight
+			shrink = inputBorderHeight
+		} else {
+			shift = 0
+			shrink = inputBorderHeight
+		}
+	}
+
+	hasListBorder := t.listBorderShape.Visible()
+	innerWidth := width
+	innerHeight := height
+	innerMarginInt := marginInt
+	innerBorderFn := func(top int, left int, width int, height int) {
+		if hasListBorder {
+			t.wborder = t.tui.NewWindow(
+				top+shift, left, width, height-shrink, tui.WindowList, tui.MakeBorderStyle(t.listBorderShape, t.unicode), false)
+		}
+	}
+	if hasListBorder {
+		if t.listBorderShape.HasTop() {
+			innerHeight--
+			innerMarginInt[0]++
+		}
+		if t.listBorderShape.HasBottom() {
+			innerHeight--
+		}
+		if t.listBorderShape.HasLeft() {
+			innerWidth -= 2
+			innerMarginInt[3] += 2
+		}
+		if t.listBorderShape.HasRight() {
+			innerWidth--
+		}
+	}
+
 	t.areaLines = height
 	t.areaColumns = width
 
@@ -1590,6 +1699,7 @@ func (t *Terminal) resizeWindows(forcePreview bool, redrawBorder bool) {
 	noBorder := tui.MakeBorderStyle(tui.BorderNone, t.unicode)
 	if forcePreview || t.needPreviewWindow() {
 		var resizePreviewWindows func(previewOpts *previewOpts)
+		stickToRight := false
 		resizePreviewWindows = func(previewOpts *previewOpts) {
 			t.activePreviewOpts = previewOpts
 			if previewOpts.size.size == 0 {
@@ -1600,7 +1710,7 @@ func (t *Terminal) resizeWindows(forcePreview bool, redrawBorder bool) {
 				pwidth := w
 				pheight := h
 				previewBorder := tui.MakeBorderStyle(previewOpts.border, t.unicode)
-				t.pborder = t.tui.NewWindow(y, x, w, h, true, previewBorder)
+				t.pborder = t.tui.NewWindow(y, x, w, h, tui.WindowPreview, previewBorder, false)
 				pwidth -= borderColumns(previewOpts.border, bw)
 				pheight -= borderLines(previewOpts.border)
 				if previewOpts.border.HasLeft() {
@@ -1615,7 +1725,7 @@ func (t *Terminal) resizeWindows(forcePreview bool, redrawBorder bool) {
 				}
 				pwidth = util.Max(0, pwidth)
 				pheight = util.Max(0, pheight)
-				t.pwindow = t.tui.NewWindow(y, x, pwidth, pheight, true, noBorder)
+				t.pwindow = t.tui.NewWindow(y, x, pwidth, pheight, tui.WindowPreview, noBorder, true)
 				if !hadPreviewWindow {
 					t.pwindow.Erase()
 				}
@@ -1645,16 +1755,25 @@ func (t *Terminal) resizeWindows(forcePreview bool, redrawBorder bool) {
 					return
 				}
 				if previewOpts.position == posUp {
+					innerBorderFn(marginInt[0]+pheight, marginInt[3], width, height-pheight)
 					t.window = t.tui.NewWindow(
-						marginInt[0]+pheight, marginInt[3], width, height-pheight, false, noBorder)
+						innerMarginInt[0]+pheight+shift, innerMarginInt[3], innerWidth, innerHeight-pheight-shrink, tui.WindowList, noBorder, true)
 					createPreviewWindow(marginInt[0], marginInt[3], width, pheight)
 				} else {
+					innerBorderFn(marginInt[0], marginInt[3], width, height-pheight)
 					t.window = t.tui.NewWindow(
-						marginInt[0], marginInt[3], width, height-pheight, false, noBorder)
+						innerMarginInt[0]+shift, innerMarginInt[3], innerWidth, innerHeight-pheight-shrink, tui.WindowList, noBorder, true)
 					createPreviewWindow(marginInt[0]+height-pheight, marginInt[3], width, pheight)
 				}
 			case posLeft, posRight:
-				pwidth := calculateSize(width, previewOpts.size, minWidth, minPreviewWidth)
+				minListWidth := minWidth
+				if t.listBorderShape.HasLeft() {
+					minListWidth += 2
+				}
+				if t.listBorderShape.HasRight() {
+					minListWidth++
+				}
+				pwidth := calculateSize(width, previewOpts.size, minListWidth, minPreviewWidth)
 				if hasThreshold && pwidth < previewOpts.threshold {
 					t.activePreviewOpts = previewOpts.alternative
 					if forcePreview {
@@ -1673,58 +1792,133 @@ func (t *Terminal) resizeWindows(forcePreview bool, redrawBorder bool) {
 				}
 				if previewOpts.position == posLeft {
 					// Put scrollbar closer to the right border for consistent look
-					if t.borderShape.HasRight() {
-						width++
+					if t.borderShape.HasRight() && !hasListBorder {
+						innerWidth++
 					}
 					// Add a 1-column margin between the preview window and the main window
+					m := 0
+					if !hasListBorder {
+						m = 1
+					}
 					t.window = t.tui.NewWindow(
-						marginInt[0], marginInt[3]+pwidth+1, width-pwidth-1, height, false, noBorder)
+						innerMarginInt[0]+shift, innerMarginInt[3]+pwidth+m, innerWidth-pwidth-m, innerHeight-shrink, tui.WindowList, noBorder, true)
 
 					// Clear characters on the margin
 					//   fzf --bind 'space:preview(seq 100)' --preview-window left,1
-					for y := 0; y < height; y++ {
-						t.window.Move(y, -1)
-						t.window.Print(" ")
+					if !hasListBorder {
+						for y := 0; y < innerHeight; y++ {
+							t.window.Move(y, -1)
+							t.window.Print(" ")
+						}
 					}
 
+					innerBorderFn(marginInt[0], marginInt[3]+pwidth, width-pwidth, height)
 					createPreviewWindow(marginInt[0], marginInt[3], pwidth, height)
 				} else {
 					// NOTE: fzf --preview 'cat {}' --preview-window border-left --border
-					if !previewOpts.border.HasRight() && t.borderShape.HasRight() {
+					stickToRight = !previewOpts.border.HasRight() && t.borderShape.HasRight()
+					if stickToRight {
+						innerWidth++
 						width++
 					}
+					innerBorderFn(marginInt[0], marginInt[3], width-pwidth, height)
 					t.window = t.tui.NewWindow(
-						marginInt[0], marginInt[3], width-pwidth, height, false, noBorder)
+						innerMarginInt[0]+shift, innerMarginInt[3], innerWidth-pwidth, innerHeight-shrink, tui.WindowList, noBorder, true)
 					x := marginInt[3] + width - pwidth
 					createPreviewWindow(marginInt[0], x, pwidth, height)
 				}
 			}
 		}
 		resizePreviewWindows(&t.previewOpts)
+
+		if t.borderShape.HasRight() && !stickToRight {
+			// Need to clear the extra margin between the borders
+			// fzf --preview 'seq 1000' --preview-window border-left --bind space:change-preview-window:border-rounded --border vertical
+			// fzf --preview 'seq 1000' --preview-window up,hidden --bind space:toggle-preview --border vertical
+			y := 0
+			if t.borderShape.HasTop() {
+				y++
+			}
+			maxY := t.border.Height()
+			if t.borderShape.HasBottom() {
+				maxY--
+			}
+			for ; y < maxY; y++ {
+				t.border.Move(y, t.border.Width()-2)
+				t.border.Print(" ")
+			}
+		}
 	} else {
 		t.activePreviewOpts = &t.previewOpts
 	}
 
 	// Without preview window
 	if t.window == nil {
-		if t.borderShape.HasRight() {
+		if t.borderShape.HasRight() && !hasListBorder {
 			// Put scrollbar closer to the right border for consistent look
+			innerWidth++
 			width++
 		}
+		innerBorderFn(marginInt[0], marginInt[3], width, height)
 		t.window = t.tui.NewWindow(
-			marginInt[0],
-			marginInt[3],
+			innerMarginInt[0]+shift,
+			innerMarginInt[3],
+			innerWidth,
+			innerHeight-shrink, tui.WindowList, noBorder, true)
+	}
+
+	// Set up input border
+	if hasInputWindow {
+		w := t.wborder
+		if t.wborder == nil {
+			w = t.window
+		}
+		if t.layout == layoutReverse {
+			t.inputBorder = t.tui.NewWindow(
+				w.Top()-shrink,
+				w.Left(),
+				w.Width(),
+				util.Min(shrink, screenHeight), tui.WindowInput, tui.MakeBorderStyle(t.inputBorderShape, t.unicode), true)
+		} else {
+			t.inputBorder = t.tui.NewWindow(
+				w.Top()+w.Height(),
+				w.Left(),
+				w.Width(),
+				util.Min(shrink, screenHeight), tui.WindowInput, tui.MakeBorderStyle(t.inputBorderShape, t.unicode), true)
+		}
+		top := t.inputBorder.Top()
+		left := t.inputBorder.Left()
+		if t.inputBorderShape.HasTop() {
+			top++
+		}
+		if t.inputBorderShape.HasLeft() {
+			left += t.borderWidth + 1
+		}
+		width := t.inputBorder.Width() - borderColumns(t.inputBorderShape, t.borderWidth)
+		if t.inputBorderShape.HasRight() {
+			width++
+		}
+		t.inputWindow = t.tui.NewWindow(
+			top,
+			left,
 			width,
-			height, false, noBorder)
+			t.inputBorder.Height()-borderLines(t.inputBorderShape),
+			tui.WindowInput, noBorder, true)
 	}
 
 	// Print border label
+	t.printLabel(t.wborder, t.listLabel, t.listLabelOpts, t.listLabelLen, t.listBorderShape, false)
 	t.printLabel(t.border, t.borderLabel, t.borderLabelOpts, t.borderLabelLen, t.borderShape, false)
 	t.printLabel(t.pborder, t.previewLabel, t.previewLabelOpts, t.previewLabelLen, t.activePreviewOpts.border, false)
+	t.printLabel(t.inputBorder, t.inputLabel, t.inputLabelOpts, t.inputLabelLen, t.inputBorderShape, false)
 }
 
 func (t *Terminal) printLabel(window tui.Window, render labelPrinter, opts labelOpts, length int, borderShape tui.BorderShape, redrawBorder bool) {
 	if window == nil {
+		return
+	}
+
+	if window.Height() == 0 {
 		return
 	}
 
@@ -1784,7 +1978,11 @@ func (t *Terminal) truncateQuery() {
 }
 
 func (t *Terminal) updatePromptOffset() ([]rune, []rune) {
-	maxWidth := util.Max(1, t.window.Width()-t.promptLen-1)
+	w := t.window
+	if t.inputWindow != nil {
+		w = t.inputWindow
+	}
+	maxWidth := util.Max(1, w.Width()-t.promptLen-1)
 
 	_, overflow := t.trimLeft(t.input[:t.cx], maxWidth)
 	minOffset := int(overflow)
@@ -1799,6 +1997,9 @@ func (t *Terminal) updatePromptOffset() ([]rune, []rune) {
 }
 
 func (t *Terminal) promptLine() int {
+	if t.inputWindow != nil {
+		return 0
+	}
 	if t.headerFirst {
 		max := t.window.Height() - 1
 		if max <= 0 { // Extremely short terminal
@@ -1813,10 +2014,25 @@ func (t *Terminal) promptLine() int {
 }
 
 func (t *Terminal) placeCursor() {
+	if t.inputWindow != nil {
+		y := t.inputWindow.Height() - 1
+		if t.layout == layoutReverse {
+			y = 0
+		}
+		t.inputWindow.Move(y, t.promptLen+t.queryLen[0])
+		return
+	}
 	t.move(t.promptLine(), t.promptLen+t.queryLen[0], false)
 }
 
 func (t *Terminal) printPrompt() {
+	w := t.window
+	if t.inputWindow != nil {
+		w = t.inputWindow
+	}
+	if w.Height() == 0 {
+		return
+	}
 	t.prompt()
 
 	before, after := t.updatePromptOffset()
@@ -1824,8 +2040,8 @@ func (t *Terminal) printPrompt() {
 	if t.paused {
 		color = tui.ColDisabled
 	}
-	t.window.CPrint(color, string(before))
-	t.window.CPrint(color, string(after))
+	w.CPrint(color, string(before))
+	w.CPrint(color, string(after))
 }
 
 func (t *Terminal) trimMessage(message string, maxWidth int) string {
@@ -1837,6 +2053,15 @@ func (t *Terminal) trimMessage(message string, maxWidth int) string {
 }
 
 func (t *Terminal) printInfo() {
+	t.withInputWindow(func() {
+		t.printInfoImpl()
+	})
+}
+
+func (t *Terminal) printInfoImpl() {
+	if t.window.Width() <= 1 || t.window.Height() == 0 {
+		return
+	}
 	pos := 0
 	line := t.promptLine()
 	maxHeight := t.window.Height()
@@ -1975,7 +2200,9 @@ func (t *Terminal) printInfo() {
 		} else {
 			outputPrinter(t.window, maxWidth-1)
 		}
-		t.window.Print(" ") // Margin
+		if fillLength >= 0 {
+			t.window.Print(" ") // Margin
+		}
 		return
 	}
 
@@ -2029,7 +2256,7 @@ func (t *Terminal) printHeader() {
 		return
 	}
 	max := t.window.Height()
-	if t.headerFirst {
+	if t.inputWindow == nil && t.headerFirst {
 		max--
 		if !t.noSeparatorLine() {
 			max--
@@ -2049,7 +2276,7 @@ func (t *Terminal) printHeader() {
 		if needReverse && idx < len(t.header0) {
 			line = len(t.header0) - idx - 1
 		}
-		if !t.headerFirst {
+		if t.inputWindow == nil && !t.headerFirst {
 			line++
 			if !t.noSeparatorLine() {
 				line++
@@ -2094,10 +2321,7 @@ func (t *Terminal) printList() {
 	count := t.merger.Length() - t.offset
 
 	// Start line
-	startLine := 2 + t.visibleHeaderLines()
-	if t.noSeparatorLine() {
-		startLine--
-	}
+	startLine := t.promptLines() + t.visibleHeaderLines()
 	maxy += startLine
 
 	barRange := [2]int{startLine + barStart, startLine + barStart + barLength}
@@ -2113,7 +2337,7 @@ func (t *Terminal) printList() {
 
 func (t *Terminal) printBar(lineNum int, forceRedraw bool, barRange [2]int) bool {
 	hasBar := lineNum >= barRange[0] && lineNum < barRange[1]
-	if hasBar != t.prevLines[lineNum].hasBar || forceRedraw {
+	if (hasBar != t.prevLines[lineNum].hasBar || forceRedraw) && t.window.Width() > 0 {
 		t.move(lineNum, t.window.Width()-1, true)
 		if len(t.scrollbar) > 0 && hasBar {
 			t.window.CPrint(tui.ColScrollbar, t.scrollbar)
@@ -2138,10 +2362,10 @@ func (t *Terminal) printItem(result Result, line int, maxLine int, index int, cu
 
 	// Avoid unnecessary redraw
 	numLines, _ := t.numItemLines(item, maxLine-line+1)
-	newLine := itemLine{firstLine: line, numLines: numLines, cy: index + t.offset, current: current, selected: selected, label: label,
+	newLine := itemLine{valid: true, firstLine: line, numLines: numLines, cy: index + t.offset, current: current, selected: selected, label: label,
 		result: result, queryLen: len(t.input), width: 0, hasBar: line >= barRange[0] && line < barRange[1]}
 	prevLine := t.prevLines[line]
-	forceRedraw := prevLine.other || prevLine.firstLine != newLine.firstLine
+	forceRedraw := !prevLine.valid || prevLine.other || prevLine.firstLine != newLine.firstLine
 	printBar := func(lineNum int, forceRedraw bool) bool {
 		return t.printBar(lineNum, forceRedraw, barRange)
 	}
@@ -2210,10 +2434,17 @@ func (t *Terminal) printItem(result Result, line int, maxLine int, index int, cu
 	}
 	if current {
 		preTask := func(marker markerClass) {
+			w := t.window.Width() - t.pointerLen
+			if w < 0 {
+				return
+			}
 			if len(label) == 0 {
 				t.window.CPrint(tui.ColCurrentCursorEmpty, t.pointerEmpty)
 			} else {
 				t.window.CPrint(tui.ColCurrentCursor, label)
+			}
+			if w-t.markerLen < 0 {
+				return
 			}
 			if selected {
 				t.window.CPrint(tui.ColCurrentMarker, markerFor(marker))
@@ -2224,10 +2455,17 @@ func (t *Terminal) printItem(result Result, line int, maxLine int, index int, cu
 		finalLineNum = t.printHighlighted(result, tui.ColCurrent, tui.ColCurrentMatch, true, true, line, maxLine, forceRedraw, preTask, postTask)
 	} else {
 		preTask := func(marker markerClass) {
+			w := t.window.Width() - t.pointerLen
+			if w < 0 {
+				return
+			}
 			if len(label) == 0 {
 				t.window.CPrint(tui.ColCursorEmpty, t.pointerEmpty)
 			} else {
 				t.window.CPrint(tui.ColCursor, label)
+			}
+			if w-t.markerLen < 0 {
+				return
 			}
 			if selected {
 				t.window.CPrint(tui.ColMarker, markerFor(marker))
@@ -2943,17 +3181,28 @@ func (t *Terminal) printAll() {
 func (t *Terminal) flush() {
 	t.placeCursor()
 	if !t.suppress {
-		windows := make([]tui.Window, 0, 4)
-		if t.borderShape != tui.BorderNone {
+		windows := make([]tui.Window, 0, 7)
+		if t.border != nil {
 			windows = append(windows, t.border)
 		}
-		if t.hasPreviewWindow() {
-			if t.pborder != nil {
-				windows = append(windows, t.pborder)
-			}
+		if t.pborder != nil {
+			windows = append(windows, t.pborder)
+		}
+		if t.pwindow != nil {
 			windows = append(windows, t.pwindow)
 		}
-		windows = append(windows, t.window)
+		if t.wborder != nil {
+			windows = append(windows, t.wborder)
+		}
+		if t.window != nil {
+			windows = append(windows, t.window)
+		}
+		if t.inputBorder != nil {
+			windows = append(windows, t.inputBorder)
+		}
+		if t.inputWindow != nil {
+			windows = append(windows, t.inputWindow)
+		}
 		t.tui.RefreshWindows(windows)
 	}
 }
@@ -3853,6 +4102,10 @@ func (t *Terminal) Loop() error {
 						if t.hasPreviewer() {
 							t.previewBox.Set(reqPreviewReady, nil)
 						}
+					case reqRedrawInputLabel:
+						t.printLabel(t.inputBorder, t.inputLabel, t.inputLabelOpts, t.inputLabelLen, t.inputBorderShape, true)
+					case reqRedrawListLabel:
+						t.printLabel(t.wborder, t.listLabel, t.listLabelOpts, t.listLabelLen, t.listBorderShape, true)
 					case reqRedrawBorderLabel:
 						t.printLabel(t.border, t.borderLabel, t.borderLabelOpts, t.borderLabelLen, t.borderShape, true)
 					case reqRedrawPreviewLabel:
@@ -3950,7 +4203,7 @@ func (t *Terminal) Loop() error {
 	previewDraggingPos := -1
 	barDragging := false
 	pbarDragging := false
-	pborderDragging := false
+	pborderDragging := -1
 	wasDown := false
 	needBarrier := true
 
@@ -4234,6 +4487,18 @@ func (t *Terminal) Loop() error {
 				} else {
 					req(reqHeader)
 				}
+			case actChangeInputLabel:
+				t.inputLabelOpts.label = a.a
+				if t.inputBorder != nil {
+					t.inputLabel, t.inputLabelLen = t.ansiLabelPrinter(a.a, &tui.ColInputLabel, false)
+					req(reqRedrawInputLabel)
+				}
+			case actChangeListLabel:
+				t.listLabelOpts.label = a.a
+				if t.wborder != nil {
+					t.listLabel, t.listLabelLen = t.ansiLabelPrinter(a.a, &tui.ColListLabel, false)
+					req(reqRedrawListLabel)
+				}
 			case actChangeBorderLabel:
 				t.borderLabelOpts.label = a.a
 				if t.border != nil {
@@ -4250,6 +4515,20 @@ func (t *Terminal) Loop() error {
 				body := t.executeCommand(a.a, false, true, true, false, "")
 				if actions, err := parseSingleActionList(strings.Trim(body, "\r\n")); err == nil {
 					return doActions(actions)
+				}
+			case actTransformInputLabel:
+				label := t.executeCommand(a.a, false, true, true, true, "")
+				t.inputLabelOpts.label = label
+				if t.inputBorder != nil {
+					t.inputLabel, t.inputLabelLen = t.ansiLabelPrinter(label, &tui.ColInputLabel, false)
+					req(reqRedrawInputLabel)
+				}
+			case actTransformListLabel:
+				label := t.executeCommand(a.a, false, true, true, true, "")
+				t.listLabelOpts.label = label
+				if t.wborder != nil {
+					t.listLabel, t.listLabelLen = t.ansiLabelPrinter(label, &tui.ColListLabel, false)
+					req(reqRedrawListLabel)
 				}
 			case actTransformBorderLabel:
 				label := t.executeCommand(a.a, false, true, true, true, "")
@@ -4639,6 +4918,14 @@ func (t *Terminal) Loop() error {
 			case actToggleWrap:
 				t.wrap = !t.wrap
 				req(reqList, reqHeader)
+			case actToggleMultiLine:
+				t.multiLine = !t.multiLine
+				req(reqList)
+			case actToggleHscroll:
+				// Force re-rendering of the list
+				t.prevLines = make([]itemLine, len(t.prevLines))
+				t.hscroll = !t.hscroll
+				req(reqList)
 			case actTrackCurrent:
 				if t.track == trackDisabled {
 					t.track = trackCurrent
@@ -4674,7 +4961,7 @@ func (t *Terminal) Loop() error {
 				if !me.Down {
 					barDragging = false
 					pbarDragging = false
-					pborderDragging = false
+					pborderDragging = -1
 					previewDraggingPos = -1
 				}
 
@@ -4730,20 +5017,36 @@ func (t *Terminal) Loop() error {
 				}
 
 				// Preview border dragging (resizing)
-				if !pborderDragging && clicked && t.hasPreviewWindow() && t.pborder.Enclose(my, mx) {
+				if pborderDragging < 0 && clicked && t.hasPreviewWindow() {
 					switch t.activePreviewOpts.position {
 					case posUp:
-						pborderDragging = my == t.pborder.Top()+t.pborder.Height()-1
+						if t.pborder.Enclose(my, mx) && my == t.pborder.Top()+t.pborder.Height()-1 {
+							pborderDragging = 0
+						} else if t.listBorderShape.HasTop() && t.wborder.Enclose(my, mx) && my == t.wborder.Top() {
+							pborderDragging = 1
+						}
 					case posDown:
-						pborderDragging = my == t.pborder.Top()
+						if t.pborder.Enclose(my, mx) && my == t.pborder.Top() {
+							pborderDragging = 0
+						} else if t.listBorderShape.HasBottom() && t.wborder.Enclose(my, mx) && my == t.wborder.Top()+t.wborder.Height()-1 {
+							pborderDragging = 1
+						}
 					case posLeft:
-						pborderDragging = mx == t.pborder.Left()+t.pborder.Width()-1
+						if t.pborder.Enclose(my, mx) && mx == t.pborder.Left()+t.pborder.Width()-1 {
+							pborderDragging = 0
+						} else if t.listBorderShape.HasLeft() && t.wborder.Enclose(my, mx) && mx == t.wborder.Left() {
+							pborderDragging = 1
+						}
 					case posRight:
-						pborderDragging = mx == t.pborder.Left()
+						if t.pborder.Enclose(my, mx) && mx == t.pborder.Left() {
+							pborderDragging = 0
+						} else if t.listBorderShape.HasRight() && t.wborder.Enclose(my, mx) && mx == t.wborder.Left()+t.wborder.Width()-1 {
+							pborderDragging = 1
+						}
 					}
 				}
 
-				if pborderDragging {
+				if pborderDragging >= 0 {
 					var newSize int
 					var prevSize int
 					switch t.activePreviewOpts.position {
@@ -4764,6 +5067,7 @@ func (t *Terminal) Loop() error {
 						offset := mx - t.pborder.Left()
 						newSize = prevSize - offset
 					}
+					newSize -= pborderDragging
 					if newSize < 1 {
 						newSize = 1
 					}
@@ -4778,6 +5082,21 @@ func (t *Terminal) Loop() error {
 					break
 				}
 
+				// Inside the input window
+				if t.inputWindow != nil && t.inputWindow.Enclose(my, mx) {
+					mx -= t.inputWindow.Left()
+					my -= t.inputWindow.Top()
+					y := t.inputWindow.Height() - 1
+					if t.layout == layoutReverse {
+						y = 0
+					}
+					mxCons := util.Constrain(mx-t.promptLen, 0, len(t.input))
+					if my == y && mxCons >= 0 {
+						t.cx = mxCons + t.xoffset
+					}
+					break
+				}
+
 				// Ignored
 				if !t.window.Enclose(my, mx) && !barDragging {
 					break
@@ -4786,10 +5105,7 @@ func (t *Terminal) Loop() error {
 				// Translate coordinates
 				mx -= t.window.Left()
 				my -= t.window.Top()
-				min := 2 + t.visibleHeaderLines()
-				if t.noSeparatorLine() {
-					min--
-				}
+				min := t.promptLines() + t.visibleHeaderLines()
 				h := t.window.Height()
 				switch t.layout {
 				case layoutDefault:
@@ -4841,7 +5157,7 @@ func (t *Terminal) Loop() error {
 
 				if me.Down {
 					mxCons := util.Constrain(mx-t.promptLen, 0, len(t.input))
-					if my == t.promptLine() && mxCons >= 0 {
+					if t.inputWindow == nil && my == t.promptLine() && mxCons >= 0 {
 						// Prompt
 						t.cx = mxCons + t.xoffset
 					} else if my >= min {
@@ -4862,7 +5178,7 @@ func (t *Terminal) Loop() error {
 						// Header
 						numLines := t.visibleHeaderLines()
 						lineOffset := 0
-						if !t.headerFirst {
+						if t.inputWindow == nil && !t.headerFirst {
 							// offset for info line
 							if t.noSeparatorLine() {
 								lineOffset = 1
@@ -5190,11 +5506,20 @@ func (t *Terminal) vset(o int) bool {
 	return t.cy == o
 }
 
-func (t *Terminal) maxItems() int {
-	max := t.window.Height() - 2 - t.visibleHeaderLines()
-	if t.noSeparatorLine() {
-		max++
+// Number of prompt lines in the list window
+func (t *Terminal) promptLines() int {
+	if t.inputWindow != nil {
+		return 0
 	}
+	if t.noSeparatorLine() {
+		return 1
+	}
+	return 2
+}
+
+// Number of item lines in the list window
+func (t *Terminal) maxItems() int {
+	max := t.window.Height() - t.visibleHeaderLines() - t.promptLines()
 	return util.Max(max, 0)
 }