+++ test/test_go.rb
@@ -1442,9 +1442,13 @@ class TestGoFZF < TestBase
     writelines(['=' * 10_000 + '0123456789'])
     [0, 3, 6].each do |off|
       tmux.prepare
-      tmux.send_keys "#{FZF} --hscroll-off=#{off} -q 0 < #{tempname}", :Enter
+      tmux.send_keys "#{FZF} --hscroll-off=#{off} -q 0 --bind space:toggle-hscroll < #{tempname}", :Enter
       tmux.until { |lines| assert lines[-3]&.end_with?((0..off).to_a.join + '··') }
       tmux.send_keys '9'
+      tmux.until { |lines| assert lines[-3]&.end_with?('789') }
+      tmux.send_keys :Space
+      tmux.until { |lines| assert lines[-3]&.end_with?('=··') }
+      tmux.send_keys :Space
       tmux.until { |lines| assert lines[-3]&.end_with?('789') }
       tmux.send_keys :Enter
     end
@@ -2133,7 +2137,11 @@ class TestGoFZF < TestBase
   end
 
   def test_keep_right
-    tmux.send_keys "seq 10000 | #{FZF} --read0 --keep-right --no-multi-line", :Enter
+    tmux.send_keys "seq 10000 | #{FZF} --read0 --keep-right --no-multi-line --bind space:toggle-multi-line", :Enter
+    tmux.until { |lines| assert lines.any_include?('9999␊10000') }
+    tmux.send_keys :Space
+    tmux.until { |lines| assert lines.any_include?('> 1') }
+    tmux.send_keys :Space
     tmux.until { |lines| assert lines.any_include?('9999␊10000') }
   end
 
@@ -2638,7 +2646,7 @@ class TestGoFZF < TestBase
   end
 
   def test_change_preview_window
-    tmux.send_keys "seq 1000 | #{FZF} --preview 'echo [[{}]]' --preview-window border-none --bind '" \
+    tmux.send_keys "seq 1000 | #{FZF} --preview 'echo [[{}]]' --no-preview-border --bind '" \
       'a:change-preview(echo __{}__),' \
       'b:change-preview-window(down)+change-preview(echo =={}==)+change-preview-window(up),' \
       'c:change-preview(),d:change-preview-window(hidden),' \
@@ -3438,6 +3446,80 @@ class TestGoFZF < TestBase
       │
       │
       │   3
+    BLOCK
+    tmux.until { assert_block(block, _1) }
+  end
+
+  def test_list_border_and_label
+    tmux.send_keys %(seq 100 | #{FZF} --border rounded --list-border double --list-label list --list-label-pos 2:bottom --header-lines 3 --query 1 --padding 1,2), :Enter
+    block = <<~BLOCK
+      │   ║   11
+      │   ║ > 10
+      │   ║   3
+      │   ║   2
+      │   ║   1
+      │   ║   19/97 ─
+      │   ║ > 1
+      │   ╚list══════
+      │
+      ╰──────────────
+    BLOCK
+    tmux.until { assert_block(block, _1) }
+  end
+
+  def test_input_border_and_label
+    tmux.send_keys %(seq 100 | #{FZF} --border rounded --input-border bold --input-label input --input-label-pos 2 --header-lines 3 --query 1 --padding 1,2), :Enter
+    block = <<~BLOCK
+      │     11
+      │   > 10
+      │     3
+      │     2
+      │     1
+      │   ┏input━━━━━
+      │   ┃   19/97
+      │   ┃ > 1
+      │   ┗━━━━━━━━━━
+      │
+      ╰──────────────
+    BLOCK
+    tmux.until { assert_block(block, _1) }
+  end
+
+  def test_list_input_border_and_label
+    tmux.send_keys %(
+      seq 100 | #{FZF} --border rounded --list-border double --input-border bold --list-label-pos 2:bottom --input-label-pos 2 --header-lines 3 --query 1 --padding 1,2 \
+      --bind 'start:transform-input-label(echo INPUT)+transform-list-label(echo LIST)' \
+      --bind 'space:change-input-label( input )+change-list-label( list )'
+    ).strip, :Enter
+    block = <<~BLOCK
+      │   ║   11
+      │   ║ > 10
+      │   ║   3
+      │   ║   2
+      │   ║   1
+      │   ╚LIST═════
+      │   ┏INPUT━━━━
+      │   ┃   19/97
+      │   ┃ > 1
+      │   ┗━━━━━━━━━
+      │
+      ╰─────────────
+    BLOCK
+    tmux.until { assert_block(block, _1) }
+    tmux.send_keys :Space
+    block = <<~BLOCK
+      │   ║   11
+      │   ║ > 10
+      │   ║   3
+      │   ║   2
+      │   ║   1
+      │   ╚ list ═══
+      │   ┏ input ━━
+      │   ┃   19/97
+      │   ┃ > 1
+      │   ┗━━━━━━━━━
+      │
+      ╰─────────────
     BLOCK
     tmux.until { assert_block(block, _1) }
   end